Move TextArea in it's own folder
