+++ src/components/text-area/index.ts
@@ -17,4 +17,4 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-export { TextInput } from './TextInput';
+export { TextArea } from './TextArea';
